Moved common methods from class InfoPane in captureGraphicsDisplay and graphicsDisplay into abstract class Pane in new module named topGraphics.
