+++ pacai/ui/captureGraphicsDisplay.py
@@ -7,8 +7,9 @@ import os
 from pacai.core.game import Directions
 from pacai.ui import graphicsUtils
 from pacai.ui import graphicsConstants
+from pacai.ui.topGraphics import Pane
 
-class InfoPane:
+class InfoPane(Pane):
     def __init__(self, layout, gridSize, redTeam, blueTeam):
         self.gridSize = gridSize
         self.width = (layout.width) * gridSize
@@ -19,20 +20,6 @@ class InfoPane:
         self.redTeam = redTeam
         self.blueTeam = blueTeam
         self.drawPane()
-
-    def toScreen(self, pos, y = None):
-        """
-        Translates a point relative from the bottom left of the info pane.
-        """
-
-        if (y is None):
-            x, y = pos
-        else:
-            x = pos
-
-        x = self.gridSize + x  # Margin
-        y = self.base + y
-        return x, y
 
     def drawPane(self):
         self.scoreText = graphicsUtils.text(self.toScreen(0, 0), self.textColor,
@@ -54,62 +41,11 @@ class InfoPane:
     def updateBlueText(self, score):
         graphicsUtils.changeText(self.blueText, self._blueScoreString())
 
-    def initializeGhostDistances(self, distances):
-        self.ghostDistanceText = []
-
-        size = 20
-        if self.width < 240:
-            size = 12
-        if self.width < 160:
-            size = 10
-
-        for i, d in enumerate(distances):
-            t = graphicsUtils.text(self.toScreen(self.width / 2 + self.width / 8 * i, 0),
-                    graphicsConstants.GHOST_COLORS[i + 1], d, "Times", size, "bold")
-            self.ghostDistanceText.append(t)
-
     def _infoString(self, score, timeleft):
         return "SCORE: % 4d, TIME: % 4d" % (score, timeleft)
 
     def updateScore(self, score, timeleft):
         graphicsUtils.changeText(self.scoreText, self._infoString(score, timeleft))
-
-    def setTeam(self, isBlue):
-        text = "RED TEAM"
-        if isBlue:
-            text = "BLUE TEAM"
-
-        self.teamText = graphicsUtils.text(self.toScreen(300, 0), self.textColor, text, "Times",
-                self.fontSize, "bold")
-
-    def updateGhostDistances(self, distances):
-        if len(distances) == 0:
-            return
-
-        if 'ghostDistanceText' not in dir(self):
-            self.initializeGhostDistances(distances)
-        else:
-            for i, d in enumerate(distances):
-                graphicsUtils.changeText(self.ghostDistanceText[i], d)
-
-    def drawGhost(self):
-        pass
-
-    def drawPacman(self):
-        pass
-
-    def drawWarning(self):
-        pass
-
-    def clearIcon(self):
-        pass
-
-    def updateMessage(self, message):
-        pass
-
-    def clearMessage(self):
-        pass
-
 
 class PacmanGraphics:
     def __init__(self, redTeam, blueTeam, zoom=1.0, frameTime=0.0, capture=False,
